Update patch set 3
Patch Set 3:

Grrr, messed of formatting. It meant to read like this:

- AOSP state: Contact picture spans room between status bar and call buttons
- State after our first change: Contact picture always stretched over the whole screen, even if a large part of it was obscured by the call buttons.
- With this patch: Contact picture during ringing is stretched to cover the space under the glowpad. In-call, the picture fills the room between top of screen and call buttons.

With this patch, the only size change of the picture during call compared to AOSP should be the inclusion of the status bar, which shouldn't make much of a difference. Please note that the picture wasn't exactly square before, as its size depends on factors like screen size and aspect ratio or the presence of a navigation bar.

Patch-set: 3