Update patch set 3
Patch Set 3:

This still doesn't fix the problem people are seeing.

"Contact picture during ringing is stretched" - during ringing is _exactly_ the time you see the picture, which is cropped and weird-looking. Here, the picture should look okay.

"In-call, the picture fills the room between top of screen and call buttons." - In-call, you will have the phone at your ear (most of the time) and the screen is off. You don't even see the picture.


So I still feel this is the wrong approach to fix this. The problem is not _during_ a call, but when you pick up your phone to answer a call.

Patch-set: 3