Update patch set 3
Patch Set 3:

Eddy, sorry, but I don't understand what your problem is. In your comments in the original patch and in the comments above you're specifically talking about in-call situations. I've fixed those, and now the problem suddenly is the incoming call screen?

I have no intentions to do any additional 'fixes' to the incoming call screen, because there actually is nothing to fix: the caller picture being shown in full screen was the whole point of the original change, so the way it's working now is the wat it's intended to work (which, judging from the review comments, a lot of people seem to agree to).

Patch-set: 3