Update patch set 3
Patch Set 3:

I did not even realise that in-call and incoming-call screens are different screens, proir to this change.

I doubt normal users know the difference as well.


This patch looks fine for the in-call screen, but the problem when you pick up your phone to call someone or to answer an incoming call remains. 

And that's the problem most people report on the forums and G+.
Enlarged and cropped pictures also decreases the quality of the picture

Patch-set: 3