Update patch set 3
Patch Set 3:

Grrr, my bad!

But if you say the smaller picture is only seen when answering a call, my previous statement still stands:

Smaller picture needs to be seen when dialing out and on incoming call, that's when you look at it. Not in-call, when the phone is at your ear.

Thanks for your efforts and response though.

Patch-set: 3