Update patch set 3
Patch Set 3:

Danny, would it not be possible to do a scaletype of centerInside and 'letterbox' with the primary/donimant color of the image? That way square images will not get cropped and the top/bottom will be filled with a complimentary color where elongated images will fit full screen?

Patch-set: 3