Update patch set 3
Patch Set 3:

And BTW, the incoming call screen isn't the only case of the contact picture being cropped: The contact details screen has a _way_ more cropped image as header (and nobody complained about that so far).

I'm thinking of shortening the transition animation a bit (from 300 to 200ms) - thoughts about that?

Patch-set: 3