Update patch set 3
Patch Set 3:

That's because in the contact details, people know it is cropped like that way, and tapping it will expand the picture. This picture is not cropped from the sides, but from top/bottom and also not blown up.
Also, people select a person themself, so they already know who is the contact.

In the incoming call screen, people experience the blown up picture as a bug. You loose details from the sides and details in quality.
Also, the incoming call screen is seen a lot during the day when people call you, where as editing a contact details is not something you do a lot.

Patch-set: 3