Update patch set 3
Patch Set 3:

In the statement 'people experience ...[this]...as a bug', you pretend that our user base is homogenous in their expectations from the phone UI. This isn't actually true, though. Some people (like you) don't like the new appearance, other people (like me and the various reviewers who gave +1 on #65023) like it. In any case, I buy the argument that users will perceive in cropped in-call photos as a bug (as the visible size of the photo didn't change much), which is why I fixed it. I fail to see why anyone would see the full-screen photo as a bug, though, and not as intentional change.

Patch-set: 3