Update patch set 1
Patch Set 1:

Well, my personal preference is on the fullscreen photo - having seen the AOSP behaviour again when doing this patch, I find it plain ugly. I guess it's just a matter of taste...

Anyway, I have no issues with making this a preference; question is just where to put it.

Patch-set: 1