Update patch set 1
Patch Set 1: Code-Review+1

Now I tested WITH this patch and USE_FULLSCREEN_CALLER_PHOTO = false and I like it much better.

Same thing happens as I stated above. when call ends before picking up, the buttons on the bottom of the screen appear. Screenhot:

https://www.dropbox.com/s/nd3o20zj25vb551/Screenshot_2014-07-20-23-00-05.png

I guess this is normal behaviour and  this time the picture does not shift like it does when using fullscreen caller photo. Much better.

I would love this to be an option, preferably the default.

Patch-set: 1