Update patch set 3
Patch Set 3:

Yes, it's enlarged a tiny bit to cover the status bar. On my Nexus 4, its size is now 768x912 pixels, compared to 768x862 pixels in AOSP. Given that the contact picture size is 720x720 pixels, the photo was cropped horizontally by 12.2% in AOSP and is cropped by 18.75% now. Does this really make that much of a difference? I can understand the dislike of the fullscreen photo cropping (size is 768x1184 there, thus 54.2% of cropping), but I don't really see the problem in the additional cropping due to the status bar.

Patch-set: 3