Update patch set 6
Patch Set 6: Code-Review-1

The record button drawable color seems a little bit off compared to the other button drawables. It's a little lighter. Can you please adjust that?

Also, wouldn't it make sense to kill the disabled drawable and - instead of enabling/disabling the record button when there is/isn't an active call - to show/hide it?

Patch-set: 6