Cleanup and fix CAF SSN handling code.
Use boolean flags in the call object instead of parsing supplementary
service notification objects. Also don't make showing a dialog for
'additional call forwarded' depend on the disconnect cause
INCOMING_MISSED, as it's not guaranteed one gets a disconnect event for
the (later forwarded) call. On mako, only the respective SSN is
delivered and assigned to the foreground call; and we want to show the
dialog when that call is finished.

Change-Id: I62828eac871453bf485e23c6455ccd85574da718