Update patch set 9
Patch Set 9:

(1 comment)

I'm getting an NPE with PS9 with CallButtonPresenter being passed a null activity.

Danny, what do you think about removing context as a parameter for CallRecorder.getInstance?  The instance of the call recorder is not really tied to any one activity context.  It must always be tied to the application context so that it can retain its service binding even if all activities are gone.  By requiring context as a parameter we open ourselves up to lifecycle bugs like this one.

If we pass the application context to call recorder when the application is created then we will always have access to the call recorder regardless of where we are in the lifecycle of the current activity.

Patch-set: 9