Update patch set 10
Patch Set 10:

Oops, that was indeed a stupid mistake.

On the topic of passing the context into getInstance(), my goal was to a) assure the call recorder object is fully initialized at any point in time and b) to decouple it from the dialer code. I realize that passing in the context is cumbersome at times though, so I've removed the parameter and initialize the CallRecorder object at the same time the InCallPresenter object is initialized.

Patch-set: 10