Update patch set 11
Patch Set 11:

The legal items have already been vetted by lawyers. 

The onus to inform parties that the call is being recorded is on the participant(s) recording, not us. 

The build time overlay prevents this from use unless the user explicitly builds with it, or in our case, the build is distributed in a region that sanctions such behavior.

Patch-set: 11