Update patch set 11
Patch Set 11: Code-Review+1

The legal burden to comply with the law is on the user, not on the supplier of the software.  Giving someone the ability to do something is not illegal, it's only illegal if they use it in such a way that does not comply with local laws.  There are quite a few extensions for skype that record calls, I don't see any of them getting sued.

This is a valuable feature that has many LEGAL uses.  So I'm going to +1 this.

Patch-set: 11