Update patch set 11
Patch Set 11:

I said onus, as in duty, not bonus. 

The obligation to inform the participants that the call is being recorded is the participant initiating the recording (where relevant). 

The build time overlay keeps this from being introduced into mass scale CM deployment (eg download.CyanogenMod.org) and we have no intention to start distributing builds with that functionality in that scale. 

As stated, this use case is for markets where all the legal items you keep bringing up are a non-factor. 

We cannot control and are not responsible for use of modified builds outside of our control. If someone wants to build this in the US and doesn't inform the other party (if required) it is _their_ fault. If they do so in the US and the tone requirements are not met, it is likewise _their_ fault. 

Our obligation starts and stops at responsible distribution of builds with this functionality - something that the build overlay already provides us.

Patch-set: 11