Update patch set 1
Patch Set 1: Code-Review-1

This will break outgoing vibration as it'll only be played if mActiveCall is null.

Besides that, IMHO state = INCALL without active call is something that should never happen. When declining calls I'd expect it to go from OUTGOING to NO_CALLS. What are the exact steps to reproduce this? I'd like to look at the state transitions.

Patch-set: 1