Update patch set 2
Patch Set 2:

I added some drawables, not entirely happy with them yet though. The record button is just a circle (and should maybe be a bit smaller?), the blacklist button is the material contact icon with a shrinked material block icon next to it (where the latter looks a bit slim due to the shrinking). Improvements are still welcome :)

Patch-set: 2