Update patch set 4
Patch Set 4:

A floppy icon is usually used for a 'save' action, I don't think it's well suited for a 'record' state (especially as we already are using the cassette recorder picture for the hold button, which has a pause icon.

I've tweaked the drawables in PS3 and am fairly happy with them now. I'll upload some screenshots tomorrow.

Patch-set: 4