Update patch set 2
Patch Set 2:

(Looking at my screenshots again, it might be possible to add a 7th button for >= 420dp and therefore a 6th button for < 420dp. Needs exploring whether this fits for hdpi though. When porting over I2865... this differentiation can be made by an overlayable integer, but this patch like won't apply cleanly due to the IMS changes in 5.1; and determination of the correct value likely needs some testing.)

Patch-set: 2