Update patch set 2
Patch Set 2:

Either that, or play around with the number of buttons to see how many of them we can reasonably fit in. I don't have an informed opinion on which of those options is better at the moment though. OTOH I'd say I like the 'compress the buttons a little further' better though, as probably looks somewhat cleaner.

Patch-set: 2