Update patch set 2
Patch Set 2:

Yeah, in calls my face will touch buttons while the screen is off due to proximity sensor.

Only happens when I add that property overlay for decoupleInteractiveModeFromDisplay. Or whatever it is. 

I enable that so that ambient display pulse won't register screen touches, but in doing so, that brings out the incallui problem without this patch.

Patch-set: 2