Update patch set 2
Patch Set 2:

It seems to not be needed. I've got no problem as long as I don't enable that one config value.

It does work fine, but I'm not sure is necessary. At least not for why I thought it was.

I can re open it to get more feedback, see what the masses want to do?

Patch-set: 2