Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > Can you provide a stack trace and some information about the device
 > you're seeing this on?

I have a faulty camera on an Xperia M that throws an assertion error when getCameraCharacteristics is invoked. Here's a logcat that illustrates this: http://pastebin.com/PdyHVKyp (Note that this logcat was captured for http://review.cyanogenmod.org/#/c/104202/, but apparently, you can see how it's related)

IMO, a bad camera shouldn't make dialer inaccessible.

I can post the stack trace for the IncallUI crash after a few days (The device is not with me now), but I don't think it's required. Catching the error at this stage seems proper IHMO. This is sort of exactly what I did with http://review.cyanogenmod.org/#/c/104202/

Patch-set: 2