Update patch set 2
Patch Set 2: Code-Review-1

IMHO this kind of assertion is something that should be handled at build time, not runtime. The problem is known, there's an obvious fix to it, so IMHO it should simply be handled correctly in the/a camera wrapper. I'm pretty sure there are a number of other devices which exhibit the same issue and have handled it in a wrapper as well.

Patch-set: 2