Update patch set 3
Patch Set 3:

> Try to fix it with a camera wrapper first.

Okay, but I'd like to mention that the reason the camera fails is not because of a software/device issue, but it's a recently incurred hardware problem. The camera no longer works even on stock ROM. (getSupportedPreviewSizes() returns null because the camera hardware does not respond properly or something). Should I still go ahead and try implementing a new camera wrapper?

Btw, Other Xperia M phones of the exact same model does *not* have this issue.

Patch-set: 3