Update patch set 2
Patch Set 2:

Fix: Can't answer or decline an inComingCall on sim1 when dialog 
mCannedResponsePopup is showing for a former InComingCall on sim2.

When this happend, mCall[] has two calls, mCall[0] is an INCOMING call,
and mCall[1] is an IDLE call,and the IDLE is obviously can not be answerd.
solution:Add a call state check for answering or decline an inComingCall,
and replace phoneId for rejectCallWithMessage().

Patch-set: 2