Update patch set 1
Patch Set 1:

Isn't better to check for ringing state. Afaik, that state is the only one that can be declined. And in the case the call coming from the second phone while the first one is in call or in hold state, the current logic will return the first phone and not the second one

Patch-set: 1