Update patch set 3
Patch Set 3:

This should work good for answer or reject an incoming call. 
But if we do so, after call disconnected with rejectCallWithMessageDialog still showing, in rejectCallWithMessage() mcall[phoneId] would be null,and DUT can't send message successfully. So in http://review.cyanogenmod.org/#/c/107674/ I save the last incomingCallPhoneId for rejectCallWithMessage().
In fact? as log as the call is disconnected, rejectCallWithMessage() can not send messages successfully, because mCall[phoneId] doesn't exsit in CallList.

Patch-set: 3