Update patch set 3
Patch Set 3:

I was playing around with openstreetmap once. The webservice that is currently used (which does not look like the official one) seems to be very slow and does not respond in time (before the lookup thread gets killed). Maybe its also dependant on your location (slow in Europe?)
So either the timeout could be increased or the implementation changed maybe (but I think the official API needs a key and stuff like that..)

Patch-set: 3