InCallUI : Don't renable keyguard if call ended and screen off
onCallListChange can get called multiple times after a call is disconnected.
Keyguard has callbacks to enable the lockscreen if the call is disconnected.
Make sure to not reset the DISMISS_KEYGUARD flag after the keyguard is enabled
since that will dismiss the lockscreen and the user will not have a lockscreen
after screen on.

Repro :

1. Accept incoming call
2. Turn screen off
3. Disconnect call in the other party
4. Turn screen on
5. Observe no lockscreen without this patch

CYNGNOS-1387

Change-Id: I1bab371293888224a1840ba04e6fdfc3fec346a4