Update patch set 3
Patch Set 3:

> > Maybe audio source related?
 > >
 > > Check here: http://review.cyanogenmod.org/#/c/130743/1/overlay/packages/apps/Dialer/res/values/config.xml
 > 
 > doc on my builds have already this cherry picked this from cm12.0,
 > although is strange because is not building the incallui apk

Same as you I already had this picket and source is set to 4, as it should be (call up and downlink). Perhaps some recent changes in the hal? I'll check ...

 > Does the audio file actually plays anything aside from static on
 > everyone's end?

Eleven just stays 'loading forever' on the PC the file plays but it's empty.

This is on titan (msm8226), we do use msm8974 audio hal.

Patch-set: 3