Update patch set 1
Patch Set 1:

FWIW, the theme was changed as part of commit 91d03c80, which introduced a SnackBar. The design library widgets mandate an AppCompat theme, which likely is was the reason for the change.

We probably could derive InCallActivity from AppCompatActivity and use getSupportActionBar() instead of getActionBar().

Patch-set: 1
Label: Code-Review=0