Update patch set 3
Patch Set 3: Code-Review+1

Note: This is a temporary solution because it is a bit of a hack. We are currently looking into creating our own "snackbar" that won't mess with other parts of the code, doesn't require coordinator layout and appcompat, and isn't a hack.

Patch-set: 3
Reviewer: Richard MacGregor <12937@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=+1