Update patch set 1
Patch Set 1:

TelephonyManager in impl is null. It's not that it's not ready, it simply is null. And the only place where that is set is when binding InCallService. I can see this happening if there is a way to start InCallActivity manually. But not knowing the intricacies of Dialer, this prevents the activity from crashing.

Patch-set: 1