Update patch set 1
Patch Set 1:

Reason:
This patch cause a bug in incoming call screen.
(In rtl,) To answer call you have to swipe left, where the button is green and the text is "?? ?????" (Do not answer).
This is bug: the text is not the action that actually performed. And also people reported it cause them to reject calls accidentaly, because they swipe right when they want to answer.
So, please revert this patch, and think how to do it without ruin that screen for RTL users.

Patch-set: 1